+++ metadata.rb
@@ -1,9 +1,9 @@
-name             'apache2-nativex'
-maintainer       'NativeX'
-maintainer_email 'adrian.herrera@nativex.com'
+name             'apache2-simplyadrian'
+maintainer       'simplyadrian'
+maintainer_email 'simplyadrian@gmail.com'
 license          'All rights reserved'
-description      'Installs/Configures apache2-nativex'
+description      'Installs/Configures apache2-simplyadrian'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
 
-depends		 "apache2"
+depends  "apache2"